backend/requirements.txt
Flask==3.1.0
python-pptx==1.0.2
google-cloud-aiplatform==1.74.0
